Done some refactoring - moved ViewModels, that were common to TimeClient and TimeServer projects into TimeProjectServices project
