+++ samples/TimeServerProject/Client/TimeClient/ViewModels/StartViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using TimeProjectServices.ViewModels;
 
 namespace TimeClient.ViewModels
 {
