+++ samples/TimeServerProject/Server/TimeServer/TimeServer.csproj
@@ -12,6 +12,7 @@
 	    <SubType>Designer</SubType>
     </AvaloniaResource>
     <AvaloniaResource Include="Assets\**" />
+    <Folder Include="Services\" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Avalonia" Version="0.10.0-preview3" />
@@ -22,5 +23,6 @@
   <ItemGroup>
     <ProjectReference Include="..\..\..\..\CustomControls\CustomControls.csproj" />
     <ProjectReference Include="..\..\..\..\NetworkingUtilities\NetworkingUtilities.csproj" />
+    <ProjectReference Include="..\..\Services\TimeProjectServices\TimeProjectServices.csproj" />
   </ItemGroup>
 </Project>
