+++ samples/TimeServerProject/Server/TimeServer/ViewModels/MainWindowViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using TimeProjectServices.ViewModels;
 
 namespace TimeServer.ViewModels
 {
