--- samples/TimeServerProject/Server/TimeServer/ViewModels/ViewModelBase.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using ReactiveUI;
-
-namespace TimeServer.ViewModels
-{
-	public class ViewModelBase : ReactiveObject
-	{
-	}
-}
