+++ samples/TimeServerProject/Services/TimeProjectServices/ViewModels/ErrorNotificationViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using Avalonia.Media;
 
-namespace TimeClient.ViewModels
+namespace TimeProjectServices.ViewModels
 {
 	internal class ErrorNotificationViewModel : NotificationViewModel
 	{
