+++ samples/TimeServerProject/Services/TimeProjectServices/ViewModels/InfoNotificationViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using Avalonia.Media;
 
-namespace TimeClient.ViewModels
+namespace TimeProjectServices.ViewModels
 {
 	internal class InfoNotificationViewModel : NotificationViewModel
 	{
