+++ samples/TimeServerProject/Services/TimeProjectServices/ViewModels/NotificationViewModelFactory.cs
@@ -1,18 +1,8 @@
 ﻿using System;
-using Avalonia.Media;
-using JetBrains.Annotations;
 using NetworkingUtilities.Utilities.Events;
 
-namespace TimeClient.ViewModels
+namespace TimeProjectServices.ViewModels
 {
-	public abstract class NotificationViewModel : ViewModelBase
-	{
-		[UsedImplicitly] public abstract Geometry Icon { get; }
-		[UsedImplicitly] public abstract string Message { get; set; }
-		[UsedImplicitly] public abstract string Title { get; }
-		public StatusCode Type { [UsedImplicitly] get; set; }
-	}
-
 	public static class NotificationViewModelFactory
 	{
 		public static NotificationViewModel Create(StatusCode code, string message) =>
